+++ pages/favorites/Favorites.tsx
@@ -1,21 +1,17 @@
 import { NavigationContainer } from '@react-navigation/native';
 import React from 'react'
-import Players from './Players';
-import Games from './components/Games';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { colors } from '../../styles/colors';
-import { TabNavIcons as GamesTabNavIcons } from './components/Games';
-import { TabNavIcons as PlayersTabNavIcons } from './Players';
-import { PageNavigationProps } from '../../navigation';
-import NavBar from '../../components/NavBar';
+import Games, { TabNavIcons as GamesTabNavIcons } from './components/Games';
+import Players, { TabNavIcons as PlayersTabNavIcons } from './components/Players';
 
 export type TabIcons = { active: string, inactive: string };
 const favoriteTabs: { [k: string]: TabIcons } = {
     'Games': GamesTabNavIcons,
     'Players': PlayersTabNavIcons,
 }
-const Favorites = ({ navigation }: PageNavigationProps<typeof RouteName>) => {
+const Favorites = () => {
     const Tab = createBottomTabNavigator();
     return (
         <>
